Added react-router-dom.
Linked the pages to the top of the NavBar (temp). Finished style of Home page and Network Status Page. Notifications Page is almost done. Settings page needs to be completed.

Layout.jsx added for continuous NavBar on all pages so it only loads once.

Files have been rearranged and are due for more